Extract HomePaga data fetch in dedicated hook
