+++ src/App.tsx
@@ -27,7 +27,7 @@ import AppHeader from './components/AppHeader'
 import { useGlobalContext } from './contexts/global'
 import AddressInfoSection from './pages/AddressInfoPage'
 import BlockInfoSection from './pages/BlockInfoPage'
-import HomeSection from './pages/HomePage'
+import HomeSection from './pages/HomePage/HomePage'
 import TransactionInfoSection from './pages/TransactionInfoPage'
 import GlobalStyle, { deviceBreakPoints } from './style/globalStyles'
 import { darkTheme, lightTheme } from './style/themes'
